// src/models/api/model/scriptEntity.ts
/**
 * Ferret Server API
 * API of Ferret Server
 *
 * OpenAPI spec version: 0.1
 * Contact: ziflex@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Entity } from './entity';
import { ScriptCommon } from './scriptCommon';
import { ScriptExecution } from './scriptExecution';
import { ScriptPersistence } from './scriptPersistence';

export interface ScriptEntity {
    id: string;
    rev: string;
    createdAt: string;
    updatedAt?: string;
    name: string;
    description?: string;
    execution: ScriptExecution;
    persistence: ScriptPersistence;
}